Add files via upload
An ACO algorithm is comprised of two main steps: build solutions and update pheromones. Procedures called artificial ants concurrently build paths (solutions)  that start at a given city, visit all other cities only once and then return to the start city. Ants decide which city will be visited next according to a stochastic policy that 
takes into account pheromone trails. 
After the solutions are built, the pheromone trails of the arcs that form the solutions are updated. The pheromone trail in a given arc decays as a result of pheromone 
evaporation, but it is made stronger by the deposition of pheromones by the ants that  traverse the arc. Arcs with higher pheromone levels tend to be chosen more often by 
ants, while arcs will lower pheromone levels tend to be chosen less.